+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix issue showing free shipping when cart has no discount, no matter the value.
+
 ## [1.1.2] - 2021-02-23
 
 ## [1.1.0] - 2021-02-04